refactor(cli): move seed script to cli module and configure entrypoint
