+++ docker-compose.prod.yml
@@ -64,7 +64,7 @@ services:
     container_name: pybot-seed
     profiles: ["seed"]
     restart: "no"
-    command: ["python", "fill_point_db.py"]
+    command: ["pybot-seed"]
 
   redis:
     image: redis:7-alpine
