+++ tests/scripts/test_fill_point_db.py
@@ -11,7 +11,7 @@ from unittest.mock import AsyncMock, Mock
 import pytest
 from sqlalchemy.ext.asyncio import AsyncSession
 
-import fill_point_db
+from pybot.cli import seed as fill_point_db
 from src.pybot.core.constants import PointsTypeEnum
 from src.pybot.dto import AdjustUserPointsDTO, CompetenceCreateDTO, CompetenceReadDTO, UserCreateDTO, UserReadDTO
 from src.pybot.dto.value_objects import Points
@@ -245,7 +245,6 @@ def test_main_uses_tyro_cli_and_asyncio_run(monkeypatch: pytest.MonkeyPatch) -> 
     )
     expected_runtime_config = fill_point_db.build_seed_config(cli_config)
     tyro_cli_mock = Mock(return_value=cli_config)
-    ensure_project_root_mock = Mock()
     observed: dict[str, object] = {}
     original_asyncio_run = asyncio.run
 
@@ -258,14 +257,12 @@ def test_main_uses_tyro_cli_and_asyncio_run(monkeypatch: pytest.MonkeyPatch) -> 
         return original_asyncio_run(coro)
 
     monkeypatch.setattr(fill_point_db.tyro, "cli", tyro_cli_mock)
-    monkeypatch.setattr(fill_point_db, "_ensure_project_root_on_path", ensure_project_root_mock)
     monkeypatch.setattr(fill_point_db, "fill_database", fake_fill_database)
     monkeypatch.setattr(fill_point_db.asyncio, "run", fake_asyncio_run)
 
     fill_point_db.main()
 
     tyro_cli_mock.assert_called_once_with(fill_point_db.FillDatabaseCLIConfig, args=None)
-    ensure_project_root_mock.assert_called_once_with()
     assert observed["is_coroutine"] is True
     assert observed["runtime_config"] == expected_runtime_config
